refactor: combine 2 graph clients
