+++ packages/fx-core/src/client/interfaces/ListSensitivityCacheValue.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export interface ListSensitivityCacheValue {
+  labels: SensitivityLabel[];
+  unixTimestamp: number;
+}
+
+export class SensitivityLabel {
+  id?: string;
+  name?: string;
+  description?: string;
+  displayName?: string;
+}
