+++ packages/fx-core/src/common/constants.ts
@@ -85,3 +85,4 @@ export const GraphTeamsChannelCreateScopes = ["Channel.Create"];
 export const GraphTeamsTeamReadScopes = ["Team.ReadBasic.All"];
 export const GraphTeamsChannelReadScopes = ["Channel.ReadBasic.All"];
 export const GraphTeamsInstallAppScopes = ["TeamsAppInstallation.ReadWriteAndConsentForTeam"];
+export const ListSensitivityLabelScope = "InformationProtectionPolicy.Read";
